all other handles added to hook
