+++ src/components/CreateDefaultLearningPath/Table/CreateDefaultLearningPathTable.hooks.tsx
@@ -1,6 +1,7 @@
+import { DragEndEvent, DragOverEvent, DragStartEvent } from '@dnd-kit/core'
+import { arrayMove } from '@dnd-kit/sortable'
 import { Dispatch, SetStateAction, useCallback, useContext, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import { CreateAlgorithmTableNameProps } from '@components'
 import { DefaultLearningPath } from '@core'
 import { SnackbarContext, postDefaultLearningPath } from '@services'
 import { usePersistedStore, useStore } from '@store'
@@ -9,18 +10,93 @@ export type useCreateDefaultLearningPathTableProps = {
   setIsSending: Dispatch<SetStateAction<boolean>>
   orderedItems: string[]
   disabledItems: string[]
+  setActiveId: (value: SetStateAction<string | null>) => void
+  setOrderedItems: Dispatch<SetStateAction<string[]>>
+  setDisabledItems: Dispatch<SetStateAction<string[]>>
 }
 
 export const useCreateDefaultLearningPathTable = ({
   setIsSending,
   orderedItems,
-  disabledItems
+  disabledItems,
+  setActiveId,
+  setOrderedItems,
+  setDisabledItems
 }: useCreateDefaultLearningPathTableProps) => {
   const { t } = useTranslation()
   const { addSnackbar } = useContext(SnackbarContext)
   const getUser = usePersistedStore((state) => state.getUser)
   const clearDefaultLearningPathCache = usePersistedStore((state) => state.clearDefaultLearningPathCache)
   const clearLearningPathElementCache = useStore((state) => state.clearLearningPathElementCache)
+
+  // Toggle disable state of an item
+  const handleToggleDisable = (itemKey: string) => {
+    if (disabledItems.includes(itemKey)) {
+      setDisabledItems((prev) => prev.filter((key) => key !== itemKey))
+    } else {
+      setDisabledItems((prev) => [...prev, itemKey])
+    }
+  }
+
+  // Reset an item from the droppable container back to the left.
+  const handleResetItem = (itemKey: string) => {
+    setOrderedItems((prev) => prev.filter((key) => key !== itemKey))
+  }
+
+  const handleRemoveAll = () => {
+    setOrderedItems([])
+    setDisabledItems([])
+  }
+
+  const handleDragStart = (event: DragStartEvent) => {
+    setActiveId(event.active.id as string)
+  }
+
+  const handleDragOver = (event: DragOverEvent) => {
+    const { active, over } = event
+    if (!active || !over) return
+    const activeIdStr = active.id as string
+    if (orderedItems.includes(activeIdStr) && over.id === 'droppable-container') {
+      const currentIndex = orderedItems.indexOf(activeIdStr)
+      const lastIndex = orderedItems.length - 1
+      if (currentIndex !== lastIndex) {
+        setOrderedItems(arrayMove(orderedItems, currentIndex, lastIndex))
+      }
+    }
+  }
+
+  const handleDragEnd = (event: DragEndEvent) => {
+    const { active, over } = event
+    setActiveId(null)
+    if (!over) return
+
+    const activeIdStr = active.id.toString()
+    const overIdStr = over.id.toString()
+
+    if (orderedItems.includes(activeIdStr)) {
+      if (overIdStr === 'droppable-container') {
+        const oldIndex = orderedItems.indexOf(activeIdStr)
+        const newIndex = orderedItems.length - 1
+        if (oldIndex !== newIndex) {
+          setOrderedItems(arrayMove(orderedItems, oldIndex, newIndex))
+        }
+      } else if (activeIdStr !== overIdStr && orderedItems.includes(overIdStr)) {
+        const oldIndex = orderedItems.indexOf(activeIdStr)
+        const newIndex = orderedItems.indexOf(overIdStr)
+        setOrderedItems(arrayMove(orderedItems, oldIndex, newIndex))
+      }
+      return
+    }
+
+    if (orderedItems.includes(overIdStr)) {
+      const newIndex = orderedItems.indexOf(overIdStr)
+      const newOrderedItems = [...orderedItems]
+      newOrderedItems.splice(newIndex, 0, activeIdStr)
+      setOrderedItems(newOrderedItems)
+    } else {
+      setOrderedItems([...orderedItems, activeIdStr])
+    }
+  }
 
   const handleSubmit = useCallback(() => {
     setIsSending(true)
@@ -69,8 +145,22 @@ export const useCreateDefaultLearningPathTable = ({
   ])
   return useMemo(
     () => ({
-      handleSubmit
+      handleSubmit,
+      handleDragEnd,
+      handleDragOver,
+      handleToggleDisable,
+      handleResetItem,
+      handleDragStart,
+      handleRemoveAll
     }),
-    [handleSubmit]
+    [
+      handleSubmit,
+      handleDragEnd,
+      handleDragOver,
+      handleToggleDisable,
+      handleResetItem,
+      handleDragStart,
+      handleRemoveAll
+    ]
   )
 }
